Feature/sessions

+++ src/main/java/fmi/eventmanager/Agendy/service/EventService.java
@@ -5,8 +5,13 @@ import fmi.eventmanager.Agendy.model.dto.Events.EventResponse;
 import fmi.eventmanager.Agendy.model.dto.Events.UpdateEventRequest;
 import fmi.eventmanager.Agendy.model.entity.Event;
 import fmi.eventmanager.Agendy.model.entity.EventStatus;
+import fmi.eventmanager.Agendy.model.entity.Role;
+import fmi.eventmanager.Agendy.model.entity.User;
 import fmi.eventmanager.Agendy.repository.EventRepository;
+import fmi.eventmanager.Agendy.repository.UserRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,12 +20,22 @@ import java.util.List;
 public class EventService {
 
     private final EventRepository eventRepository;
+    private final UserRepository userRepository;
 
-    public EventService(EventRepository eventRepository) {
+    public EventService(EventRepository eventRepository, UserRepository userRepository) {
         this.eventRepository = eventRepository;
+        this.userRepository = userRepository;
     }
 
     public EventResponse createEvent(CreateEventRequest request, Long organizerId) {
+
+        User user = userRepository.findById(organizerId)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found"));
+
+        if (user.getRole() != Role.ADMIN) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Only admins can add events!");
+        }
+
         Event event = new Event(
                 organizerId,
                 request.getTitle(),
@@ -49,10 +64,14 @@ public class EventService {
         return responses;
     }
 
-    public EventResponse updateEvent(Long id, UpdateEventRequest request) {
+    public EventResponse updateEvent(Long id, UpdateEventRequest request, Long userId) {
 
         Event event = eventRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Event not found"));
+
+        if (!event.getOrganizerId().equals(userId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not the organizer of this event");
+        }
 
         event.setTitle(request.getTitle());
         event.setVenue(request.getVenue());
@@ -62,9 +81,14 @@ public class EventService {
         return mapToResponse(saved);
     }
 
-    public void deleteEvent(Long id) {
-        eventRepository.findById(id)
+    public void deleteEvent(Long id, Long userId) {
+        Event event = eventRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Event not found"));
+
+        if (!event.getOrganizerId().equals(userId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not the organizer of this event");
+        }
+
         eventRepository.deleteById(id);
     }
